requirements.txt
aiohttp>=3.8.0
python-socketio>=5.1.0
aiohttp-cors>=0.7.0
speechmatics-python>=2.1.0
sounddevice>=0.4.6
numpy>=1.20.0
python-dotenv>=0.19.0
websockets>=10.0